1.硬币器识别的问题，将找零器按照硬币器上报。即STATUS_RPT中CC_ST的状态即为找零器状态(正常、禁用、故障、不存在)。目前的做法是改动了隐藏维护菜单里的设置，而且不是出厂默认设置，希望新版本中予以优化。      2、找零器硬币余量的上报。目前的版本中是将找零器中1元硬币余量50或者0作为STATUS_RPT中change1余量上报，这样在友宝后台看来就是5角余量50，1元余量为0。新版本软件中请将1元硬币余量50或者0写入到change2余量中，change1余量填0，这样后台看起来更直观，方便加盟用户运营管理。

+++ API/ChangerAPI.c
@@ -171,9 +171,9 @@ void Hopper_Init()
 ** output parameters:   无
 ** Returned value:      1：配币成功；0：配备失败
 *********************************************************************************************************/
-uint8_t Hopper_Distribute(uint32_t nMoney,uint8_t *tHp1,uint8_t *tHp2,uint8_t *tHp3)
+uint8_t Hopper_Distribute(uint32_t nMoney,uint32_t *tHp1,uint32_t *tHp2,uint32_t *tHp3)
 {
-	uint8_t Count = 0;
+	uint32_t Count = 0;
 	uint32_t i,nMoney1=0,nMoney2=0;
 	uint8_t ln1=0,ln2=0,ln3=0;
 
@@ -265,14 +265,14 @@ uint8_t Hopper_Distribute(uint32_t nMoney,uint8_t *tHp1,uint8_t *tHp2,uint8_t *t
 *********************************************************************************************************/
 uint32_t Hopper_Dispence(uint32_t ChangeMoney,uint32_t *SurplusMoney,uint8_t *Hp1OutNum,uint8_t *Hp2OutNum,uint8_t *Hp3OutNum)
 {
-	uint8_t Hp1Out=0,Hp2Out=0,Hp3Out=0;
+	uint32_t Hp1Out=0,Hp2Out=0,Hp3Out=0;
 	uint8_t res,mErrFlag = 5,err;
 	uint8_t ln1=0,ln2=0,ln3=0,i;
 	MessagePack *HpBackMsg;
 	uint8_t hpoutcount = 0;
 	uint32_t LastMoney = 0;
 
-	//确定三个Hopper的单位面值大小顺序
+	//1.确定三个Hopper的单位面值大小顺序
 	for(i=0;i<3;i++)
 	{
 		if(stEvbHp[i].Num == 1)
@@ -290,11 +290,15 @@ uint32_t Hopper_Dispence(uint32_t ChangeMoney,uint32_t *SurplusMoney,uint8_t *Hp
 			ln3 = i;
 		}
 	}
+	//2.如果找零金额小于最小面值，则直接退出
 	if(ChangeMoney < stEvbHp[ln1].Price)
 	{
 		*SurplusMoney = ChangeMoney; 
 		return 0;
 	}
+
+
+	
 	while(mErrFlag)
 	{
 		//对Hopper进行配币
